Move delete_instance_health_checks to a separate module
