+++ src/tests/_internal/server/background/scheduled_tasks/test_instances.py
@@ -40,7 +40,6 @@ from dstack._internal.core.models.runs import (
     JobStatus,
 )
 from dstack._internal.server.background.scheduled_tasks.instances import (
-    delete_instance_health_checks,
     process_instances,
 )
 from dstack._internal.server.models import (
@@ -65,7 +64,6 @@ from dstack._internal.server.testing.common import (
     ComputeMockSpec,
     create_fleet,
     create_instance,
-    create_instance_health_check,
     create_job,
     create_project,
     create_repo,
@@ -1205,39 +1203,6 @@ class TestAddSSHInstance:
         assert instance.status == InstanceStatus.IDLE
         assert instance.total_blocks == expected_blocks
         assert instance.busy_blocks == 0
-
-
-@pytest.mark.asyncio
-@pytest.mark.parametrize("test_db", ["sqlite", "postgres"], indirect=True)
-@pytest.mark.usefixtures("test_db", "image_config_mock")
-class TestDeleteInstanceHealthChecks:
-    async def test_deletes_instance_health_checks(
-        self, monkeypatch: pytest.MonkeyPatch, session: AsyncSession
-    ):
-        project = await create_project(session=session)
-        instance = await create_instance(
-            session=session, project=project, status=InstanceStatus.IDLE
-        )
-        # 30 minutes
-        monkeypatch.setattr(
-            "dstack._internal.server.settings.SERVER_INSTANCE_HEALTH_TTL_SECONDS", 1800
-        )
-        now = get_current_datetime()
-        # old check
-        await create_instance_health_check(
-            session=session, instance=instance, collected_at=now - dt.timedelta(minutes=40)
-        )
-        # recent check
-        check = await create_instance_health_check(
-            session=session, instance=instance, collected_at=now - dt.timedelta(minutes=20)
-        )
-
-        await delete_instance_health_checks()
-
-        res = await session.execute(select(InstanceHealthCheckModel))
-        all_checks = res.scalars().all()
-        assert len(all_checks) == 1
-        assert all_checks[0] == check
 
 
 @pytest.mark.asyncio
